backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
python-dotenv
textblob
textstat
nltk
django-redis
redis
